pyproject.toml
[tool.poetry]
name = "ops2deb"
version = "0.4.1"
description = "Build debian packages"
repository = "https://github.com/upciti/ops2deb"
license = "MIT"
keywords=["Debian", "DevOps", "Automation", "Kubernetes", "APT"]
authors = ["Upciti <support@upciti.com>"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Code Generators",
    "Typing :: Typed",
]
readme = "README.md"
packages = [
    {include = "ops2deb", from  = "src"},
]

[tool.poetry.dependencies]
python = "^3.8"
httpx = ">=0.15.4,<0.19.0"
typer = "*"
pydantic = "*"
aiofiles = "*"
semver = "3.0.0-dev.2"
jinja2 = "*"
"ruamel.yaml" = "^0.17.16"

[tool.poetry.scripts]
ops2deb = "ops2deb.cli:main"

[tool.poetry.dev-dependencies]
starlette = "*"
pytest = "*"
mypy = "*"
taskipy = "*"
pytest-cov = "*"
flake8 = "*"
black = ">=20.8b1"
isort = "^5.8.0"
safety = "*"

[tool.taskipy.tasks]
check = """
isort src tests\
 && black .\
 && flake8 src\
 && mypy src\
 && pytest --cov src tests\
"""

[tool.black]
line-length = 90
target_version = ['py38']
exclude = '''
(
  /(
    | \.git
    | \.venv
    | \.mypy_cache
    | src/ops2deb/templates
  )/
)
'''

[tool.isort]
line_length = 90
profile = "black"

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
